backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.39.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
fastapi==0.136.1
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.13
isodate==0.7.2
msrest==0.7.1
oauthlib==3.3.1
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
